test: 코치 가능 시간, 예약 목록, 정보 api 문서 최신화
