+++ backend/src/docs/asciidoc/api.adoc
@@ -29,15 +29,15 @@
 
 == 면담 가능 시간 API
 === [코치] 면담 가능 시간 저장 API
-operation::member-controller-test/save-calendar-times[snippets='curl-request,http-request,http-response']
+operation::coach-controller-test/save-calendar-times[snippets='http-request,http-response']
 
 === [코치] 면담 가능 시간 조회 API
-operation::member-controller-test/find-calendar-times[snippets='curl-request,http-request,http-response']
+operation::coach-controller-test/find-calendar-times[snippets='http-request,http-response']
 
 
 == 면담 예약 API
 === [코치] 면담 예약 목록 조회 API
-operation::coach-controller-test/find-all-by-coach[snippets='curl-request,http-request,http-response']
+operation::coach-controller-test/find-all-reservation-by-coach[snippets='http-request,http-response']
 
 === [크루] 코치 목록 조회 API
 operation::member-controller-test/find-coaches[snippets='curl-request,http-request,http-response']
@@ -59,3 +59,10 @@ operation::reservation-controller-test/cancel-reservation[snippets='http-request
 
 === [크루] 면담 예약 취소 API
 operation::reservation-controller-test/delete-reservation[snippets='curl-request,http-request,http-response']
+
+== 회원 API
+=== [코치] 내 정보 조회 API
+operation::coach-controller-test/find-coach[snippets='http-request,http-response']
+
+=== [코치] 내 정보 수정 API
+operation::coach-controller-test/update-coach[snippets='http-request,http-response']
