+++ backend/src/test/java/com/woowacourse/ternoko/api/MemberControllerTest.java
@@ -3,7 +3,6 @@ package com.woowacourse.ternoko.api;
 import static com.woowacourse.ternoko.config.AuthorizationExtractor.AUTHORIZATION;
 import static com.woowacourse.ternoko.config.AuthorizationExtractor.BEARER_TYPE;
 import static com.woowacourse.ternoko.fixture.CoachAvailableTimeFixture.MONTH_REQUEST;
-import static com.woowacourse.ternoko.fixture.CoachAvailableTimeFixture.NOW_MONTH_REQUEST;
 import static com.woowacourse.ternoko.fixture.MemberFixture.COACH1;
 import static com.woowacourse.ternoko.fixture.MemberFixture.COACH3;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -35,36 +34,6 @@ class MemberControllerTest extends RestDocsTestSupport {
         mockMvc.perform(MockMvcRequestBuilders
                         .get("/api/login/check")
                         .queryParam("nickname", COACH3.getNickname()))
-                .andExpect(status().isOk())
-                .andDo(restDocs.document());
-    }
-
-    @Test
-    @DisplayName("코치의 면담 가능 시간을 저장한다.")
-    void saveCalendarTimes() throws Exception {
-        mockMvc.perform(MockMvcRequestBuilders
-                        .put("/api/calendar/times")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .characterEncoding("utf-8")
-                        .header(AUTHORIZATION, BEARER_TYPE + jwtProvider.createToken(String.valueOf(COACH1.getId())))
-                        .content(objectMapper.writeValueAsString(MONTH_REQUEST)))
-                .andExpect(status().isOk())
-                .andDo(restDocs.document());
-    }
-
-    @Test
-    @DisplayName("코치의 면담 가능 시간 목록을 조회한다.")
-    void findCalendarTimes() throws Exception {
-        // given
-        createCalendarTimes(COACH1.getId());
-
-        // when, then
-        mockMvc.perform(MockMvcRequestBuilders
-                        .get("/api/calendar/times")
-                        .header(AUTHORIZATION, BEARER_TYPE + jwtProvider.createToken(String.valueOf(COACH1.getId())))
-                        .queryParam("coachId", String.valueOf(COACH1.getId()))
-                        .queryParam("year", String.valueOf(NOW_MONTH_REQUEST.getYear()))
-                        .queryParam("month", String.valueOf(NOW_MONTH_REQUEST.getMonth())))
                 .andExpect(status().isOk())
                 .andDo(restDocs.document());
     }
